Practical tuning of classifiers
This project entails practical tuning of classifiers using cross-validation and a hold-out test set. The Naive Bayes classifier is used, and it is trained in Scikit-learn. The standard one is the Gaussian variant, which assumes the data has a normal distribution. At the end of the project, the objective is to compare and select the best classifier with the highest accuracy.