Convolution Neural Network using CIFAR10 dataset
This project demonstrates training a simple Convolution Neural Network (CNN) to classify CIFAR images with the use of Keras Sequential API. The CIFAR10 dataset contains 60,000 color images in 10 classes, with 6,000 images in each class. The dataset is divided into 50,000 training images and 10,000 testing images. The classes are mutually exclusive and there is no overlap between them. The sizes of the test and training set is verified, before compiling and training the model.